ДЗ спринт 21 (v.0)

+++ app/src/main/java/com/practicum/playlistmaker/data/dto/TrackGetRequest.kt
@@ -1,5 +1,5 @@
 package com.practicum.playlistmaker.data.dto
 
 data class TrackGetRequest(
-    val trackId: Int
+    val trackId: Long
 )
