+++ app/src/main/java/com/practicum/playlistmaker/domain/search/SearchInteractor.kt
@@ -1,6 +1,6 @@
 package com.practicum.playlistmaker.domain.search
 
-import com.practicum.playlistmaker.domain.models.Track
+import com.practicum.playlistmaker.domain.model.Track
 import kotlinx.coroutines.flow.Flow
 
 interface SearchInteractor {
@@ -9,5 +9,5 @@ interface SearchInteractor {
     fun addTrackToHistory(track: Track)
 
     fun getTracks(term: String): Flow<Pair<List<Track>?, String?>>
-    fun getOneTrack(trackId: Int): Flow<Pair<List<Track>?, String?>>
+    fun getOneTrack(trackId: Long): Flow<Pair<List<Track>?, String?>>
 }
